+++ src/components/List/VirtualizedListItems/VirtualizedListItems.module.scss
@@ -1,0 +1,8 @@
+@import "../../../styles/states.scss";
+
+.scrollableContainer {
+  margin: 0;
+  padding: 0;
+  overflow: auto;
+  @include scroller();
+}
